Add GraphQL queries to components
I seem to have run into a bug with the Header component. It's
not receiving anything from the GraphQL query. I checked and this
is apparently a known issue with Gatsby, although it was supposed
to have been fixed already. I tried the suggested steps, no luck.

I'm pretty demotivated now. So I'm going to call it a day.