Issue #16, Make header responsive
I made a nav menu for smaller screens. I'm not convinced that's
the way I want to go, and I'm trying to think of
alternatives, but it works for now. The menu items have gap
between them that I need to figure out how to remove, and the
text is off center. The biggest issue is that I need to figure
out how to close the menut when the user clicks outside of it.