Fix the gap around nav menu items
At least part of the problem came from the default index.css that
comes with the gatsby starter. I haven't removed it completely
because for the most part it seems okay to me, but if it continues
to be a problem I may start over.
There are still the issues in chrome, I expect they may also
come from the default index.css. I'll check that later.