Create seperate templates for Project & Blog pages
This one is rather a large commit, but it's all related.
I want to be able to write both my project pages and
my blog posts as markdown files, at first I tried adding
a conditional operator to the resolve function in gatsby-node.js
to return only the relvent fields for each page. It turns out
that doesn't work, which makes sense given my understanding of
how Gatsby works, but it did take me a while to realize.

Instead I added a switch statement to the (formerly) BlogPostTemplate
so that it returns different markup based on the type of the post.