Add rss.xml with gatsby-plugin-feed
I renamed the "posts" directory to "blog" to match the slug in
the recommended configuration.

This one was a headache, the default configuration requires a
slug be generated for each markdown file but this wasn't made
clear in the documentation, rather it was implied that nothing
was needed besides adding the plugin to gatsby-config.js.

I found the solution in the GatsbyJS issues on GitHub. It's supposed
to be added to the docs, as I understand it. So hopefully others
won't have the same issue.