Add tag pages
I followed the official Gatsby tutorial for adding a tag page, which
for some reason uses lodash even though none of the others seem to.
After having a working setup I refactored lodash out, not that I
have anything against it per se but I didn't want to import it just
for this.